+++ packages/response/set_headers.ts
@@ -1,33 +1,35 @@
 import { headerEntries } from "./header_entries.ts";
 
 /**
- * Append the given headers to the given Response.
+ * Set the given headers in the given Response.
  *
  * The headers will be merged directly into the headers of the Response
  * if the Response headers are mutable, otherwise a new Response will be created.
  *
  * @param res The response
- * @param headers The additional headers to be appended to the response headers
+ * @param headers The additional headers to be set in the response headers
  */
-export function appendHeaders(
+export function setHeaders(
   res: Response,
   headers: HeadersInit,
 ): Response {
   const entries = headerEntries(headers);
 
-  function appendToHeaders(headers: Headers) {
+  function setToHeaders(headers: Headers) {
     for (const [name, value] of entries) {
-      headers.append(name, value);
+      if (headers.get(name) !== value) {
+        headers.set(name, value);
+      }
     }
   }
 
   if (entries.length) {
     try {
-      appendToHeaders(res.headers);
+      setToHeaders(res.headers);
     } catch {
       // Response headers were probably read-only, so create a new Response
       const headers = new Headers(res.headers);
-      appendToHeaders(headers);
+      setToHeaders(headers);
       res = new Response(res.body, {
         status: res.status,
         statusText: res.statusText,